+++ src/Farmer/Common.fs
@@ -4081,6 +4081,7 @@ module ContainerApp =
 
     type ServiceBusScaleRule = {
         QueueName: string
+        Namespace: string
         MessageCount: int
         SecretRef: string
     }